The Beginning I am writing my own website for managing RPG campaigns. It is supposed to make it easier for Dungeon Masters and Player to keep track of rules and action happening on the virtual board.
In first commit I add basic html, style and drag.js.
I started from finding a way to make divs draggable and placable in any position,
because I thought it's gonna be the most difficult thing to implement.
Turns out I took me less than half an hour, because I found
a post on stackoverflow with someone, who had the same problem.
https://stackoverflow.com/questions/17992543/how-do-i-drag-an-image-smoothly-around-the-screen-using-pure-javascript

After studying the code and implementing it, I modified it, so that an image
can not be dragged outside of #map div.

My next step is to create map functionalities:
- multiple scenes / switching maps
- locking and unlocking moving of certain 'player tokens'
- impassable objects, e.g. walls

Later I will implement more useful features, when I'm done with basic functionalities for the map.